🚚 Move connect out of nft book route
