+++ src/util/firebase.ts
@@ -61,7 +61,7 @@ export const likeNFTFreeMintTxCollection = getCollectionIfDefined(
   FIRESTORE_NFT_FREE_MINT_TX_ROOT,
 );
 export const likeNFTBookCollection = getCollectionIfDefined(FIRESTORE_LIKER_NFT_BOOK_ROOT);
-export const likeNFTBookUserCollection = getCollectionIfDefined(
+export const likeNFTConnectedUserCollection = getCollectionIfDefined(
   FIRESTORE_LIKER_NFT_BOOK_USER_ROOT,
 );
 export const likeButtonUrlCollection = getCollectionIfDefined(FIRESTORE_LIKE_URL_ROOT);
