+++ test/stub/util/firebase.ts
@@ -219,7 +219,7 @@ export const likeNFTSubscriptionUserCollection = createCollection([]);
 export const likeNFTSubscriptionTxCollection = createCollection([]);
 export const likeNFTFreeMintTxCollection = createCollection([]);
 export const likeNFTBookCollection = createCollection([]);
-export const likeNFTBookUserCollection = createCollection([]);
+export const likeNFTConnectedUserCollection = createCollection([]);
 export const likeButtonUrlCollection = createCollection([]);
 export const iscnInfoCollection = createCollection([]);
 export const iscnMappingCollection = createCollection([]);
